Crate github_readme_copy

github_readme_copy

Copy my public README.md files from Github in HTML format
version: 1.0.31 date: 2022-05-14 author: bestia.dev repository: Github

Motivation

It looks like google search SEO is really bad for github README.md files.
Maybe it will work better as HTML files on my own domain. Yes, it does.
On every README HTML, there is a link to bestia.dev and a link to the github repository.
I need a utility CLI that copies the README files as they are rendered on github and save them as html files in the directory copied_readme.
Then I will use an rsync command to upload the files to my google cloud virtual machine.

Octocrab

I will use the crate octocrab to get a list of my public repos.
You need to have a github PAT (personal access token) and save it in a environment variable:

export GITHUB_TOKEN=ghp_111111111111111111111

TODO

rayon to get html in parallel. upload to web server

cargo crev reviews and advisory

We live in times of danger with supply chain attacks.

It is recommended to always use cargo-crev to verify the trustworthiness of each of your dependencies.

Please, spread this info.

You can also read crev reviews quickly on the web:

https://web.crev.dev/rust-reviews/crates/

open-source and free as a beer

My open-source projects are free as a beer (MIT license).

I just love programming.

But I need also to drink. If you find my projects and tutorials helpful,please buy me a beer donating on my paypal.

You know the price of a beer in your local bar ;-) So I can drink a free beer for your health :-)

Na zdravje! Alla salute! Prost! Nazdravlje! 🍻

Modules

github_mod

All the real code is inside modules in separate files.

utils_mod

various utilities

Enums

LibraryError

all possible library errors for thiserror

Functions

format_upper_hello_phrase

format the hello phrase with uppercase name if it is already uppercase, return error with thiserror

list_my_public_repos

list_my_public_repos