github_readme_copy
Module encrypt_decrypt_with_ssh_key_mod

Modules

encrypt_decrypt_mod

Generic functions to encrypt and decrypt secrets using the ssh private key.

github_api_token_with_oauth2_mod

decrypt github api token from file or use the oauth2 device workflow to get the access token and encrypt it and save into file

Constants

BLUE

ANSI color

GREEN

ANSI color

RED

ANSI color

RESET

ANSI color

YELLOW

ANSI color