github_readme_copy

Module utils_mod

various utilities

Traits

ResultLogError

Trait to log the error from Result before propagation with ?.

Functions

find_from

find from_pos

find_pos_end_data_before_delimiter

return the position of end of the delimited data before the delimiter

find_pos_start_data_after_delimiter

return the position of start of the delimited data after the delimiter

tracing_init

Initialize tracing to file logs/github_readme_copy.log. \